For some reason, on Ubuntu 14.04, we have an old 'losetup' so it uses kpartx instead; unfortunately kpartx does not appear to be able to delete a mapping, leaving a stale mapping in place.
This results in the raspbian .img file being essentially empty (only
partition information is present, but no filesystem contents.)

Turns out we can use losetup -d to remove the mapping but we need to
also remove the mapping first with kpartx before using losetup -d.. This
was discovered with trial/error and I don't have a real explanation.

Sigh.